Policy-consumer/Policy-consumer/Policy-consumer/src/main/java/com/cognizant/pas/consumer/exception/GlobalExceptionHandler.java
package com.cognizant.pas.consumer.exception;

import java.time.LocalDateTime;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.RestControllerAdvice;
import org.springframework.web.context.request.WebRequest;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

import lombok.extern.slf4j.Slf4j;

@RestControllerAdvice
@Slf4j
public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
    
    /**
     * 
     * @param ex
     * @return
     */
    @ExceptionHandler(AuthorizationException.class)
    public ResponseEntity<ExceptionDetails> handleAuthorizationException(AuthorizationException ex){
        ExceptionDetails exceptionDetail = new ExceptionDetails(LocalDateTime.now(), HttpStatus.FORBIDDEN, ex.getMessage());
        log.error(ex.getMessage());
        return new ResponseEntity<>(exceptionDetail, HttpStatus.FORBIDDEN);
    }
    
    /***
     * 
     * @param ex
     * @return
     */
    @ExceptionHandler(BusinessPropertyNotFoundException.class)
    public ResponseEntity<ExceptionDetails> handleBusinessPropertyNotFoundException(BusinessPropertyNotFoundException ex){
        ExceptionDetails exceptionDetail = new ExceptionDetails(LocalDateTime.now(), HttpStatus.NOT_FOUND, ex.getMessage());
        log.error(ex.getMessage());
        return new ResponseEntity<>(exceptionDetail, HttpStatus.NOT_FOUND);
    }
    
    /***
     * 
     * @param ex
     * @return
     */
    @ExceptionHandler(ConsumerBusinessNotFoundException.class)
    public ResponseEntity<ExceptionDetails> handleConsumerBusinessNotFoundExceptionn(ConsumerBusinessNotFoundException ex){
        ExceptionDetails exceptionDetail = new ExceptionDetails(LocalDateTime.now(), HttpStatus.NOT_FOUND, ex.getMessage());
        log.error(ex.getMessage());
        return new ResponseEntity<>(exceptionDetail, HttpStatus.NOT_FOUND);
    }
    

    
    @ExceptionHandler(Exception.class)
    public ResponseEntity<Object> handleGlobalException(Exception ex, WebRequest request) {
        ExceptionDetails exceptionDetail = new ExceptionDetails(LocalDateTime.now(), HttpStatus.UNAUTHORIZED, ex.getMessage());
        log.error(ex.getMessage());
        return new ResponseEntity<>(exceptionDetail, HttpStatus.UNAUTHORIZED);
    }
    
}